Adding a template name into handlebars that way the error is slightly better
