+++ test_resources/file-with-error
@@ -1,0 +1,3 @@
+{
+    "error": "{{keepass-2-file "something"}}"
+}
